fix: Fix bugs in AVLTree and its test
In the previous version, the AVLTree unittest tested
BinarySearchTree insead of AVLTree. This bug has been
fixed after this version.

In the previous version, the attribute "key" was declared
private in BinarySearchTree. This resulted in AVLTree not
being able to access "key", when AVLTree inherited BinarySearchTree.
The attribute "key" has been declared protected.